fixed the quote calculation. it was first not calculating nothing but now it works

+++ CarInsurance/CarInsurance/ViewModels/AdminVm.cs
@@ -10,6 +10,6 @@ namespace CarInsurance.ViewModels
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string EmailAddress { get; set; }
-        public double Quote { get; set; }
+        public decimal Quote { get; set; }
     }
 }
